refactor: break the meta cited-records import cycle and extract the free-form USER-turn builder
