+++ src/synthorg/meta/chief_of_staff/_chat_format.py
@@ -1,17 +1,42 @@
 # module-kind: code
-"""Pure prompt-context formatters for the Chief of Staff chat.
+"""Prompt-context formatters and USER-turn assembly for the CoS chat.
 
 Extracted from ``chat.py`` so the service module stays well under its
-size budget. Each helper renders a readable text block from typed
-inputs; the caller in ``chat.py`` owns the untrusted-content fencing
-around any block that carries human- or agent-authored fields.
+size budget. The ``format_*`` helpers each render a readable text block
+from typed inputs; ``render_free_form_user`` assembles the full,
+fully-fenced free-form USER message (reading recent outcomes and
+applying the SEC-1 ``<task-data>`` fencing itself, since it owns the
+whole turn rather than a single block).
 """
+
+from typing import Final
 
 from synthorg.budget.currency import format_cost
 from synthorg.core.approval import ApprovalItem
+from synthorg.core.critical_errors import reraise_critical
 from synthorg.core.types import NotBlankStr
-from synthorg.meta.chief_of_staff.org_state import OrgStateSnapshot
+from synthorg.engine.prompt_safety import TAG_TASK_DATA, wrap_untrusted
+from synthorg.meta.chief_of_staff.models import ChatQuery
+from synthorg.meta.chief_of_staff.org_state import OrgStateSnapshot, format_org_state
+from synthorg.meta.chief_of_staff.prompts import CHAT_QUERY_USER
+from synthorg.meta.chief_of_staff.protocol import OutcomeStore
 from synthorg.meta.models import OrgSignalSnapshot
+from synthorg.observability import get_logger, safe_error_description
+from synthorg.observability.events.chief_of_staff import COS_CHAT_FAILED
+
+logger = get_logger(__name__)
+
+# Rendered into the prompt when the org-state read model could not be built
+# (persistence disconnected or the approval store unwired). System-authored,
+# so it is NOT fenced as untrusted content; it instructs the model to admit
+# it cannot see task / project / approval state rather than infer idleness.
+# Deliberately names no specific missing dependency: the operator-facing
+# cause (which subsystem is unwired) is carried by the WARNING the request
+# helper logs, not baked into the user-facing answer where it could be wrong.
+_ORG_STATE_UNAVAILABLE: Final[str] = (
+    "The org-state read model is currently unavailable, so I cannot see"
+    " task, project, or approval state."
+)
 
 
 def free_form_sources(
@@ -116,3 +141,60 @@ def format_signal_context(ctx: dict[str, object]) -> str:
         Resulting string.
     """
     return "\n".join(f"{k}: {v}" for k, v in ctx.items())
+
+
+async def render_free_form_user(
+    *,
+    outcome_store: OutcomeStore | None,
+    query: ChatQuery,
+    snapshot: OrgSignalSnapshot,
+    scoped_proposal: ApprovalItem | None,
+    org_state: OrgStateSnapshot | None,
+) -> str:
+    """Render the fenced USER message for a free-form question.
+
+    Reads recent outcomes for context (degrading to a placeholder on a
+    store read failure) and folds a resolved ``scoped_proposal`` summary
+    ahead of them, folds the real org-state block (or the "cannot see
+    state" sentinel), then fences every attacker-controllable field in a
+    ``<task-data>`` envelope. The org-state records are human/agent-authored,
+    so the rendered block is fenced; the unavailable sentinel is
+    system-authored and stays unfenced.
+
+    Returns:
+        The rendered, fully fenced USER-role message.
+    """
+    recent_context = "No recent proposals or alerts."
+    if outcome_store is not None:
+        try:
+            recent = await outcome_store.recent_outcomes(limit=5)
+        except Exception as exc:  # noqa: BLE001 -- criticals re-raised
+            reraise_critical(exc)
+            # A graceful degrade (falls back to a placeholder), so WARNING
+            # not ERROR; carry the redacted error context the sibling
+            # provider-failure logs also emit, so the failure is diagnosable.
+            logger.warning(
+                COS_CHAT_FAILED,
+                reason="outcome_store_read_failed",
+                error_type=type(exc).__name__,
+                error=safe_error_description(exc),
+            )
+            recent = ()
+        if recent:
+            lines = [
+                f"- {o.title} ({o.decision}, {o.decided_at:%Y-%m-%d})" for o in recent
+            ]
+            recent_context = "Recent outcomes:\n" + "\n".join(lines)
+    if scoped_proposal is not None:
+        scoped = format_scoped_proposal(scoped_proposal)
+        recent_context = f"{scoped}\n\n{recent_context}"
+    if org_state is not None:
+        org_state_block = wrap_untrusted(TAG_TASK_DATA, format_org_state(org_state))
+    else:
+        org_state_block = _ORG_STATE_UNAVAILABLE
+    return CHAT_QUERY_USER.format(
+        snapshot_summary=wrap_untrusted(TAG_TASK_DATA, format_snapshot(snapshot)),
+        org_state=org_state_block,
+        recent_context=wrap_untrusted(TAG_TASK_DATA, recent_context),
+        user_question=wrap_untrusted(TAG_TASK_DATA, query.question),
+    )
